code optimized, seperate class created for legend related code
